factor out admire toggling
