+++ apps/mobile/src/components/Feed/Socialize/AdmireButton.tsx
@@ -1,16 +1,10 @@
-import { useCallback } from 'react';
 import { TouchableOpacityProps } from 'react-native';
-import { trigger } from 'react-native-haptic-feedback';
-import { ConnectionHandler, graphql, useFragment } from 'react-relay';
-import { SelectorStoreUpdater } from 'relay-runtime';
+import { graphql, useFragment } from 'react-relay';
+import { useToggleAdmire } from 'src/hooks/useToggleAdmire';
 
 import { GalleryTouchableOpacity } from '~/components/GalleryTouchableOpacity';
 import { AdmireButtonFragment$key } from '~/generated/AdmireButtonFragment.graphql';
-import { AdmireButtonMutation } from '~/generated/AdmireButtonMutation.graphql';
 import { AdmireButtonQueryFragment$key } from '~/generated/AdmireButtonQueryFragment.graphql';
-import { AdmireButtonRemoveMutation } from '~/generated/AdmireButtonRemoveMutation.graphql';
-import { AdditionalContext, useReportError } from '~/shared/contexts/ErrorReportingContext';
-import { usePromisifiedMutation } from '~/shared/relay/usePromisifiedMutation';
 
 import { AdmireIcon } from './AdmireIcon';
 
@@ -24,13 +18,7 @@ export function AdmireButton({ eventRef, queryRef, style }: Props) {
   const event = useFragment(
     graphql`
       fragment AdmireButtonFragment on FeedEvent {
-        id
-        dbid
-
-        viewerAdmire {
-          id
-          dbid
-        }
+        ...useToggleAdmireFragment
       }
     `,
     eventRef
@@ -39,225 +27,20 @@ export function AdmireButton({ eventRef, queryRef, style }: Props) {
   const query = useFragment(
     graphql`
       fragment AdmireButtonQueryFragment on Query {
-        viewer {
-          __typename
-
-          ... on Viewer {
-            user {
-              id
-              dbid
-              username
-            }
-          }
-        }
+        ...useToggleAdmireQueryFragment
       }
     `,
     queryRef
   );
 
-  const reportError = useReportError();
-
-  const [admire] = usePromisifiedMutation<AdmireButtonMutation>(graphql`
-    mutation AdmireButtonMutation($eventId: DBID!, $connections: [ID!]!) @raw_response_type {
-      admireFeedEvent(feedEventId: $eventId) {
-        ... on AdmireFeedEventPayload {
-          __typename
-          feedEvent {
-            viewerAdmire
-              @appendNode(edgeTypeName: "FeedEventAdmireEdge", connections: $connections) {
-              dbid
-              __typename
-              creationTime
-              admirer {
-                id
-                dbid
-                username
-              }
-            }
-          }
-        }
-        ... on ErrAdmireAlreadyExists {
-          __typename
-        }
-      }
-    }
-  `);
-
-  const [removeAdmire] = usePromisifiedMutation<AdmireButtonRemoveMutation>(graphql`
-    mutation AdmireButtonRemoveMutation($admireId: DBID!) @raw_response_type {
-      removeAdmire(admireId: $admireId) {
-        ... on RemoveAdmirePayload {
-          __typename
-          admireID
-        }
-
-        ... on ErrAdmireNotFound {
-          __typename
-        }
-      }
-    }
-  `);
-
-  const interactionsConnection = ConnectionHandler.getConnectionID(
-    event.id,
-    'Interactions_admires'
-  );
-  const notesModalConnection = ConnectionHandler.getConnectionID(
-    event.id,
-    'NotesList_interactions'
-  );
-
-  const handleRemoveAdmire = useCallback(async () => {
-    trigger('impactLight');
-
-    if (!event.viewerAdmire?.dbid) {
-      return;
-    }
-
-    const errorMetadata: AdditionalContext['tags'] = {
-      eventId: event.dbid,
-    };
-
-    const updater: SelectorStoreUpdater<AdmireButtonRemoveMutation['response']> = (
-      store,
-      response
-    ) => {
-      if (response?.removeAdmire?.__typename === 'RemoveAdmirePayload') {
-        const pageInfo = store.get(interactionsConnection)?.getLinkedRecord('pageInfo');
-
-        pageInfo?.setValue(((pageInfo?.getValue('total') as number) ?? 1) - 1, 'total');
-
-        if (response.removeAdmire.admireID) {
-          const relayId = `Admire:${response.removeAdmire.admireID}`;
-
-          store.delete(relayId);
-        }
-      }
-    };
-
-    try {
-      const response = await removeAdmire({
-        updater,
-        optimisticUpdater: updater,
-        optimisticResponse: {
-          removeAdmire: {
-            __typename: 'RemoveAdmirePayload',
-            admireID: event.viewerAdmire.dbid,
-          },
-        },
-        variables: {
-          admireId: event.viewerAdmire.dbid,
-        },
-      });
-
-      if (
-        response.removeAdmire?.__typename !== 'RemoveAdmirePayload' &&
-        // We can silently fail if the post was already not admired
-        response.removeAdmire?.__typename !== 'ErrAdmireNotFound'
-      ) {
-        reportError(
-          `Could not unadmire feed event, typename was ${response.removeAdmire?.__typename}`,
-          {
-            tags: errorMetadata,
-          }
-        );
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        reportError(error);
-      } else {
-        reportError(`Could not remove admire on feed event for an unknown reason`, {
-          tags: errorMetadata,
-        });
-      }
-    }
-  }, [event.dbid, event.viewerAdmire?.dbid, interactionsConnection, removeAdmire, reportError]);
-
-  const handleAdmire = useCallback(async () => {
-    trigger('impactLight');
-
-    if (query.viewer?.__typename !== 'Viewer') {
-      return;
-    }
-
-    const errorMetadata: AdditionalContext['tags'] = {
-      eventId: event.dbid,
-    };
-
-    const updater: SelectorStoreUpdater<AdmireButtonMutation['response']> = (store, response) => {
-      if (response?.admireFeedEvent?.__typename === 'AdmireFeedEventPayload') {
-        const pageInfo = store.get(interactionsConnection)?.getLinkedRecord('pageInfo');
-
-        pageInfo?.setValue(((pageInfo?.getValue('total') as number) ?? 0) + 1, 'total');
-      }
-    };
-
-    try {
-      const optimisticAdmireId = Math.random().toString();
-      const response = await admire({
-        updater,
-        optimisticUpdater: updater,
-        optimisticResponse: {
-          admireFeedEvent: {
-            __typename: 'AdmireFeedEventPayload',
-            feedEvent: {
-              id: event.id,
-              viewerAdmire: {
-                __typename: 'Admire',
-                id: `Admire:${optimisticAdmireId}`,
-                dbid: optimisticAdmireId,
-                creationTime: new Date().toISOString(),
-                admirer: {
-                  id: query.viewer?.user?.id ?? 'unknown',
-                  dbid: query.viewer?.user?.dbid ?? 'unknown',
-                  username: query.viewer?.user?.username ?? null,
-                },
-              },
-            },
-          },
-        },
-        variables: {
-          eventId: event.dbid,
-          connections: [interactionsConnection, notesModalConnection],
-        },
-      });
-
-      if (
-        response.admireFeedEvent?.__typename !== 'AdmireFeedEventPayload' &&
-        // We can silently fail if the post was already admired
-        response.admireFeedEvent?.__typename !== 'ErrAdmireAlreadyExists'
-      ) {
-        reportError(
-          `Could not admire feed event, typename was ${response.admireFeedEvent?.__typename}`,
-          {
-            tags: errorMetadata,
-          }
-        );
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        reportError(error);
-      } else {
-        reportError(`Could not post comment on feed event for an unknown reason`, {
-          tags: errorMetadata,
-        });
-      }
-    }
-  }, [
-    admire,
-    event.dbid,
-    event.id,
-    interactionsConnection,
-    notesModalConnection,
-    query.viewer,
-    reportError,
-  ]);
-
-  const hasViewerAdmiredEvent = Boolean(event.viewerAdmire);
+  const { toggleAdmire, hasViewerAdmiredEvent } = useToggleAdmire({
+    eventRef: event,
+    queryRef: query,
+  });
 
   return (
     <GalleryTouchableOpacity
-      onPress={hasViewerAdmiredEvent ? handleRemoveAdmire : handleAdmire}
+      onPress={toggleAdmire}
       className="flex justify-center align-center w-8 h-8 pt-0.5"
       style={style}
       eventElementId="Admire Button"
